NFC: if region2 set don't disable NFC on phone shutdown
If NFC is not disabled from the UI, the user should be able to continue
making transactions. The NFC controller cannot process transactions with
the disable pin set high. The region 2 command configures the NFC
controller to opperate when the phone is off. Thus we don't disable the
NFC controller after issuing the region 2 command.

Change-Id: I610e1a6ee3b326d95c4794b01b38f2e21de85ba5
Acked-by: Umesh Jagga <ujagga@qti.qualcomm.com>
Signed-off-by: Satish Singh <ssing@codeaurora.org>