NFC: NFC reset wait causes bark due to excessive logging
It takes almost 15 msec for reset notification to be received from
NFCC and due to this NFC kernel device was waiting for 10uSec loop
which was causing bark due to excessive logging.
GPIO was also not needed to be controlled twice. Add some logging
for the NFC device for debugging.

Change-Id: I4335cbff61f16a2d93fccea8c7e39f0b430e0009
CRs-Fixed: 653737
Signed-off-by: Houston Hoffman <hhoffman@codeaurora.org>